spodkast_event_manager/requirements.txt
functions-framework
requests
flask
pydub
google-cloud-pubsub
openai
gcsfs